VERSION 0.7.0 options -e and -w
On ZFS-FUSE renameat2(.., RENAME_NOCLOBBER) fails.
Because this comes a bit unexpected, there now is a workaround
to move files using the (previously used) rename().

Option -e suppresses this fact, so you can enforce safe mode
on filesystems you know that they can be safely be used.

Option -w also enables RENAME_WHITEOUT which implies -e.
I am not sure that this should be the default in future.