VERSION 0.7.3: minor things, better cmpanddel.sh
cmpanddel.sh "protected" the directory to remove.
This moved the src directory away for a fraction of a second,
which did a race condition at my side.

As only empty directories are removed, there is no need do
"protect" it like with files and so on,
because if anything changes on the src directory,
the "rmdir" will fail anyway.  Hence this is safe to do directly.